Data acquisition

+++ py_drone_toast.py
@@ -59,15 +59,12 @@ from flask import render_template, Response, redirect, url_for
 from flask_cors import CORS
 from jinja2.exceptions import TemplateNotFound
 
-# thread Imports
-from threading import Thread
-from queue import Queue
-
 # LANDRS imports
 import graph.py_drone_graph as ldg
-import py_drone_mavlink
 from config.config_generate_form import generate_form
 from config.config_form2rdf import Form2RDFController
+from data_acquisition import data_acquisition
+from data_acquisition.data_acquisition import Data_acquisition
 
 # Defines ######################################################################
 # things I need to know
@@ -140,7 +137,7 @@ if graph_location:
         if os.path.isfile(config_file_dynamic):
             print("Loading dynamic ini")
             config.read(config_file_dynamic)
-    else:
+    elif os.path.isfile(config_file_dynamic):
         # else remove it
         os.remove(config_file_dynamic)
     
@@ -188,11 +185,11 @@ if 'GRAPH' in config.keys():
     # get dictionary
     graph_dict = config['GRAPH']
 
-# get mavlink dictionary
-mavlink_dict = {}
-if 'MAVLINK' in config.keys():
+# get data acquisition dictionary
+dataacquisition_dict = {}
+if 'DATAACQUISITION' in config.keys():
     # get dictionary
-    mavlink_dict = config['MAVLINK']
+    dataacquisition_dict = config['DATAACQUISITION']
 
 # load the data to serve on the API ############################################
 '''
@@ -215,21 +212,16 @@ if 'FLIGHT' in config.keys():
     # get dictionary
     flight_dict = config['FLIGHT']
 
-#get port. here as sent to mavlink##############################################
+#get port. here as sent to data acquisition#####################################
 port = int(get_config('DEFAULT', 'port', '5000'))
 
 ################################################################################
-# start mavlink thread, start as daemon so terminates with the main program
+# start data acquisition thread, start as daemon so terminates with the main program
+# now created in data_acquistion class
 ################################################################################
-# create queue
-q_to_mavlink = Queue()
 
-# create thread for mavlink link, send api callback
-t1 = Thread(target=py_drone_mavlink.mavlink, daemon=True,
-            args=(q_to_mavlink, mavlink_dict, 'http://localhost:' + str(port) + '/api/v1/store'))
-
-# Start mavlink thread
-t1.start()
+# create datalogger class
+data_acquire = Data_acquisition(dataacquisition_dict, 'http://localhost:' + str(port) + '/api/v1/store')
 
 ################################################################################
 # Main Flask program to provide API for drone interface
@@ -277,7 +269,7 @@ def home():
     return json.dumps(op_dict), 200, {'Content-Type': 'application/sparql-results+json; charset=utf-8'}
 
 ####################################################
-# control mavlink thread
+# control data acquisition thread
 ####################################################
 
 
@@ -304,15 +296,14 @@ def mavlink():
             ret.update( {'graphs': graph_list} )
 
     # if thread is running then send the payload
-    if t1.is_alive():
-        # status message
-        response = "message sent: " + response
+    # status message
+    response = "message sent: " + response
 
-        # message to thread
-        q_to_mavlink.put(request_dict)
+    # message to thread
+    data_acquire.q_to_data_acqu_put(request_dict)
 
     # update return
-    ret.update({"status": response, "thread": t1.is_alive()})
+    ret.update({"status": response})
 
     # go
     return ret, 200, {'Content-Type': 'application/sparql-results+json; charset=utf-8'}
@@ -572,8 +563,18 @@ def form():
 
     # show comm ports?
     comms_ports = []
-    if mavlink_dict.get('list_ports', 'False') == 'True':
-        comms_ports = py_drone_mavlink.get_serial_ports()
+    if dataacquisition_dict.get('list_ports', 'False') == 'True':
+        comms_ports = data_acquisition.get_serial_ports()
+
+        # original port?
+        if comms_ports:
+            orig_port = dataacquisition_dict.get('address', None)
+            if orig_port:
+                comms_ports.append(orig_port)
+
+            # list ports
+            for cp in comms_ports:
+                print("COMM Port", cp)
 
     # drone name?
     drone_name = config.get('DRONE', 'name', fallback=None)
@@ -657,7 +658,7 @@ def drone():
 
 @app.route('/drone_config', methods=['POST'])
 def drone_config():
-    # get request as dict to send to mavlink
+    # get request as dict to send to data acquisition
     request_dict = request.form.to_dict()
 
     # parse input form and create drone node
@@ -735,7 +736,7 @@ def flight():
 
 @app.route('/flight_create', methods=['POST'])
 def flight_create():
-    # get request as dict to send to mavlink
+    # get request as dict to send to data acquisition
     request_dict = request.form.to_dict()
     #print("REQ", request_dict)
 
@@ -760,21 +761,21 @@ def flight_create():
                 config_dyn.read(config_file_dynamic)
 
             # setup config file
-            config_dyn['MAVLINK'] = {'observation_collection': obs_col, 'dataset': dataset}
-            config.set('MAVLINK', 'observation_collection', obs_col)
-            config.set('MAVLINK', 'dataset', dataset)
+            config_dyn['DATAACQUISITION'] = {'observation_collection': obs_col, 'dataset': dataset}
+            config.set('DATAACQUISITION', 'observation_collection', obs_col)
+            config.set('DATAACQUISITION', 'dataset', dataset)
 
             # remove old sensor data
             prop_label = flight_dict.get('flight_sensor', 'sensor')
-            k_remove = [key for key, val in mavlink_dict.items() if prop_label == key[:len(prop_label)]]
+            k_remove = [key for key, val in dataacquisition_dict.items() if prop_label == key[:len(prop_label)]]
             for kr in k_remove:
-                mavlink_dict.pop(kr)
+                dataacquisition_dict.pop(kr)
 
             # load sensor data
             for sensor in mission_dict['sensors']:
                 for k in sensor:
-                    config.set('MAVLINK', k, sensor[k])
-                    config_dyn.set('MAVLINK', k, sensor[k])
+                    config.set('DATAACQUISITION', k, sensor[k])
+                    config_dyn.set('DATAACQUISITION', k, sensor[k])
 
             config_dyn['FLIGHT'] = {'flight': flt_name}
             config.set('FLIGHT', 'flight', flt_name)
@@ -783,14 +784,10 @@ def flight_create():
             with open(config_file_dynamic, 'w') as configfile:
                 config_dyn.write(configfile)
 
-           # mavlink running? if its not alive, start
-            if not t1.is_alive():
-                t1.start()
-
             # message to thread
             request_dict = {'action': 'set_oc_sensor', 'observation_collection': obs_col, 
                             'sensors': mission_dict['sensors'], 'dataset': dataset}
-            q_to_mavlink.put(request_dict)
+            data_acquire.q_to_data_acqu_put(request_dict)
 
             # create return success alert
             alert_popup = 'Flight created,\nFlight name: \t\t' + mission_dict['flight'] + \
